Rewrite class firstTest after review, deleted some useless files
